Rewrite store runtime for SQLite
Replace the remaining runtime PGlite/Postgres store paths with the SQLite facade. The root DB, chat DBs, action DBs, root migrations, chat schema, message lookup, memories, reminders, usage tracking, media-to-text cache, HD media updates, and workspace repositories now use SQLite-compatible DDL, timestamp comparisons, JSON extraction, and JSON text storage.

Update tests and action fixtures to seed SQLite-compatible data instead of Postgres JSONB, vector, and full-text expressions. Keep the legacy PGlite cluster maintenance scanner unchanged so old-data diagnostics can still inspect prior stores.

Verification: pnpm type-check; pnpm test (1020 tests, 207 suites).

+++ tests/e2e-adapter.test.js
@@ -230,7 +230,7 @@ describe("message filtering", () => {
       testUserResponseRegistry,
     );
 
-    const { rows } = await testDb.sql`SELECT COUNT(*)::int AS count FROM messages WHERE chat_id = ${chatId}`;
+    const { rows } = await testDb.sql`SELECT COUNT(*) AS count FROM messages WHERE chat_id = ${chatId}`;
     const messageCount = /** @type {{ count: number }} */ (rows[0]).count;
 
     assert.equal(messageCount, 0, "Ignored messages should not be persisted");
@@ -264,7 +264,7 @@ describe("message filtering", () => {
       },
     );
 
-    const { rows } = await testDb.sql`SELECT COUNT(*)::int AS count FROM messages WHERE chat_id = ${chatId}`;
+    const { rows } = await testDb.sql`SELECT COUNT(*) AS count FROM messages WHERE chat_id = ${chatId}`;
     const messageCount = /** @type {{ count: number }} */ (rows[0]).count;
 
     assert.equal(downloadCalled, false, "Ignored captions should short-circuit before media download");
